Problem: namespace getting very confusing
Solution: seperate namespaces into normal, user, and moderator for all views and logic
