+++ lib/routes.js
@@ -1,9 +1,6 @@
 import { FlowRouter } from 'meteor/kadira:flow-router';
 import { BlazeLayout } from 'meteor/kadira:blaze-layout';
 
-
-
-
 FlowRouter.route('/', {
   name: 'desktop',
   // subscriptions: function() {
@@ -30,9 +27,15 @@ FlowRouter.route( '/currency/:_id', {
   }
 });
 
-FlowRouter.route( '/pending', {
+FlowRouter.route( '/mypending', {
   action: function( params, queryParams ) {
-    BlazeLayout.render('editAnything', { main: 'pendingCurrencies'});
+    BlazeLayout.render('editAnything', { main: 'userPendingCurrencies'});
+  }
+});
+
+FlowRouter.route( '/moderator', {
+  action: function( params, queryParams ) {
+    BlazeLayout.render('editAnything', { main: 'moderatorDash'});
   }
 });
