+++ views/layouts/editAnything.html
@@ -7,7 +7,6 @@
     <div class="col-xs-12 col-sm-12 col-md-10 col-lg-8 col-xl-8" style="margin-bottom:10px;">
         {{> Template.dynamic template=main}}
     </div>
-
 </div>
 </div>
 </template>
