+++ views/templates/maintainerDash.html
@@ -1,0 +1,5 @@
+<template name="moderatorDash">
+  {{#each currencies}}
+      {{> moderatorPendingCurrency}}
+  {{/each}}
+</template>
