+++ views/templates/normal/currency.html
