+++ views/templates/normal/currencyDetail.html
