+++ views/templates/normal/returnedCurrencies.html
