+++ views/templates/users/addCoin.html
