+++ views/templates/users/userPendingCurrency.html
@@ -1,8 +1,8 @@
-<template name="pendingCurrency">
+<template name="userPendingCurrency">
   <div class="card border-secondary mb-3 mb-3" style="min-width:300px;overflow:hidden;">
   <div class="card-header" style="height:50px"><h4 class="pull-left">{{currencyName}} ({{currencySymbol}})</h4></div>
   <div class="card-body">
-    Status: pending maintainer approval...
+    Status: pending moderator approval...
   </div>
   </div>
 </template>
